Converted package baea version BLP_LIB_BDE_BAE_2.5.0 from clearcase

+++ groups/bae/baea/baea_performancemonitor.h
@@ -683,7 +683,7 @@ int
 baea_PerformanceMonitor::numRegisteredPids() const
 {
     bcemt_ReadLockGuard<bcemt_RWMutex> guard(&d_mapGuard);
-    return d_pidMap.size();
+    return static_cast<int>(d_pidMap.size());
 }
 
 }  // close namespace BloombergLP
